Update iTunes version parsing to match mhbd format
The ipod linux document shows an offset of 16 for iTunes version in the mhbd object, which seems to reasonably line up with the existing iTunesDB samples and the new one I pulled from my iPod. Mine was written recently with MacOS Finder/Apple Music syncing. It seems like the versions are more recent than ipod linux's lates documented version (iTunes 7.4), which makes sense since that was released in 2007.

Looking at Wikipedia, iTunes went up to version 12.9.5.5 (2019) on MacOS and continues on Windows. I'm not sure if there is any open documentation of how to line up the u32 in the iTunesDB file, but I though my file was broken when I saw "N/A (xxx)" so I thought adding something more specific would be valuable.