Treat timestamps as optional.
It seems like the timestamp fields are often 0. In some cases, this
seems to clearly indicate a null value, e.g. an unplayed song has no
"last played at" timestamp. In other cases, it may just be missing
metadata, but I think it's generally valid to accept that timestamps may
be unset.