Sample iTunesDB generated by gnupod.
I'm currently using gnupod to manage songs on my ipod. It seems to
adhere roughly to the latest documented format from ipod linux, but has
some quirks. So far, I've noticed that the mhbd language field is not
set. I haven't looked any further than that, but I thought that it might
be an interesting version of the iTunesDB file to test against.